requirements.txt
google-adk
google-generativeai
python-dotenv
beautifulsoup4
requests
